Move to legacy namespace
